A Major refactor to store all information with each server instance.
The idea here, is to allow configuring multiple nnhelplightning backends
in gnus, so that you can interact with multiple workspaces (or servers)
at the same time.

However, this does not year appear to be working as expected with
multiple backends. GNUS sees to be correctly associating groups with the
correct server (if you press ^ to go into server view, then enter a
server, you only see the groups tied to it), but clicking on a group
always passes in the first server, not the server associated with that
group.

I am still investigating multiple instances. However, this change still
works fine with a single instance, and cleans things up.